Safari-609.1.11.1/LayoutTests/html5lib/generated/run-tests11-write-expected.txt

CONSOLE MESSAGE: line 168: Warning: Problem parsing viewBox=""
CONSOLE MESSAGE: line 168: Warning: Problem parsing viewBox=""
CONSOLE MESSAGE: line 168: Warning: Problem parsing viewBox=""
../resources/tests11.dat: PASS
